requirements.txt
accelerate==1.1.1
aiofiles==23.2.1
aiohappyeyeballs==2.4.3
aiohttp==3.11.7
aiosignal==1.3.1
albucore==0.0.13
albumentations==1.4.10
annotated-types==0.7.0
anyio==3.7.1
astor==0.8.1
attrs==24.2.0
azure-core==1.32.0
azure-identity==1.19.0
beautifulsoup4==4.12.3
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
click==8.1.7
contourpy==1.3.1
cryptography==43.0.3
cycler==0.12.1
Cython==3.0.11
decorator==5.1.1
defusedxml==0.7.1
dill==0.3.9
distro==1.9.0
easyocr==1.7.2
einops==0.8.0
evdev==1.7.1
fastapi==0.115.5
ffmpy==0.4.0
filelock==3.16.1
fire==0.7.0
fonttools==4.55.0
frozenlist==1.5.0
fsspec==2024.10.0
gradio==5.6.0
gradio_client==1.4.3
grandalf==0.8
greenlet==3.1.1
h11==0.14.0
httpcore==1.0.7
httpx==0.27.2
httpx-sse==0.4.0
huggingface-hub==0.26.2
idna==3.10
imageio==2.36.0
imgaug==0.4.0
Jinja2==3.1.4
jiter==0.7.1
joblib==1.4.2
jsonpatch==1.33
jsonpointer==3.0.0
kiwisolver==1.4.7
langchain==0.3.7
langchain-core==0.3.20
langchain-openai==0.2.9
langchain-text-splitters==0.3.2
langgraph==0.2.53
langgraph-checkpoint==2.0.5
langgraph-sdk==0.1.36
langsmith==0.1.145
lazy_loader==0.4
lmdb==1.5.1
lxml==5.3.0
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.9.2
mdurl==0.1.2
MouseInfo==0.1.3
mpmath==1.3.0
msal==1.31.1
msal-extensions==1.2.0
msgpack==1.1.0
mss==10.0.0
multidict==6.1.0
networkx==3.4.2
ninja==1.11.1.2
numpy==1.26.4
nvidia-cublas-cu12==12.4.5.8
nvidia-cuda-cupti-cu12==12.4.127
nvidia-cuda-nvrtc-cu12==12.4.127
nvidia-cuda-runtime-cu12==12.4.127
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.2.1.3
nvidia-curand-cu12==10.3.5.147
nvidia-cusolver-cu12==11.6.1.9
nvidia-cusparse-cu12==12.3.1.170
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.4.127
openai==1.55.0
opencv-contrib-python==4.10.0.84
opencv-python==4.10.0.84
opencv-python-headless==4.10.0.84
opt-einsum==3.3.0
orjson==3.10.11
packaging==24.2
paddleocr==2.9.1
paddlepaddle==2.6.2
pandas==2.2.3
pillow==11.0.0
playsound==1.3.0
portalocker==2.10.1
propcache==0.2.0
protobuf==5.28.3
psutil==6.1.0
pvporcupine==3.0.3
py-cpuinfo==9.0.0
PyAutoGUI==0.9.54
pyclipper==1.3.0.post6
pycparser==2.22
pydantic==2.9.2
pydantic_core==2.23.4
pydub==0.25.1
PyGetWindow==0.0.9
Pygments==2.18.0
PyJWT==2.10.0
PyMsgBox==1.0.9
pyparsing==3.2.0
pyperclip==1.9.0
PyRect==0.2.0
PyScreeze==1.0.1
python-bidi==0.6.3
python-dateutil==2.9.0.post0
python-docx==1.1.2
python-multipart==0.0.12
python3-xlib==0.15
pytweening==1.2.0
pytz==2024.2
PyYAML==6.0.2
RapidFuzz==3.10.1
regex==2024.11.6
requests==2.32.3
requests-toolbelt==1.0.0
rich==13.9.4
ruff==0.8.0
safehttpx==0.1.1
safetensors==0.4.5
scikit-image==0.24.0
scikit-learn==1.5.2
scipy==1.14.1
seaborn==0.13.2
semantic-version==2.10.0
shapely==2.0.6
shellingham==1.5.4
six==1.16.0
sniffio==1.3.1
sounddevice==0.5.1
soupsieve==2.6
SQLAlchemy==2.0.36
starlette==0.41.3
supervision==0.18.0
sympy==1.13.1
tenacity==9.0.0
termcolor==2.5.0
thop==0.1.1.post2209072238
threadpoolctl==3.5.0
tifffile==2024.9.20
tiktoken==0.8.0
timm==1.0.11
tokenizers==0.20.3
tomli==2.1.0
tomlkit==0.12.0
torch==2.5.1
torchvision==0.20.1
tqdm==4.67.0
transformers==4.46.3
triton==3.1.0
typer==0.13.1
typing_extensions==4.12.2
tzdata==2024.2
ultralytics==8.1.24
urllib3==2.2.3
uvicorn==0.32.1
watchdog==6.0.0
webrtcvad==2.0.10
websockets==12.0
yarl==1.18.0
python-dotenv==1.0.1